MMIO Peripherals based on templates with AMP Demo
We require the ability to describe a region of memory as an MMIO
peripheral, probably based on a template, while allowing default
behaviour for the remaining region (e.g. logging).

Furthermore we require instances of this template to be dispatched to
the peripheral model server and for receives to return and write to
memory.

To achieve this we use AvataterPeripheral as our base device and add a
special template decorator that extracts out the peripheral name.

Signed-off-by: Antony Vennard <antony@vennard.ch>